Update WordUtils
This PR updates all usages of WordUtils from the deprecated
org.apache.commons.lang3.text package (deprecated since commons-lang3
v3.6) to the recommended org.apache.commons.text package. The change is
necessary as the project is using commons-lang3 v3.13.0, and the
deprecated API should be replaced with the supported version.

- Updates 20 Java files to import WordUtils from org.apache.commons.text
  instead of the deprecated org.apache.commons.lang3.text package
- Ensures compatibility with the currently used commons-lang3 v3.13.0
  library
- Leverages existing commons-text-1.10.0.jar already present in the
  server and client sub-projects and adds the same library to the donkey
  sub-project

Signed-off-by: Richard Ogin <rogin@users.noreply.github.com>